fix : 스크롤 바 크기 6px로 수정
