+++ frontend/src/components/organisms/ReplyBar/styles.ts
@@ -1,8 +1,9 @@
 import Label from '@atoms/Label';
 import ImageButton from '@atoms/ImageButton';
-import ThreadContent from '@molecules/ThreadContent';
+
 import BrowseChannelHeader from '@molecules/BrowseChannelHeader';
 import styled from 'styled-components';
+
 const Container = styled.div<{ widthVW: number; isOpened: boolean }>`
   display: ${({ isOpened }) => (isOpened ? 'flex' : 'none')};
   flex-direction: column;
@@ -17,10 +18,6 @@ export const StyledBrowseChannelHeader = styled(BrowseChannelHeader)`
   z-index: 5;
   --saf-0: rgba(var(--sk_foreground_low, 29, 28, 29), 0.13);
   box-shadow: 0 0 0 1px var(--saf-0), 0 4px 12px 0 rgba(0, 0, 0, 0.12);
-`;
-
-export const StyledThreadContent = styled(ThreadContent)`
-  margin-top: 25px;
 `;
 
 export const StyledLabel = styled(Label)`
