+++ frontend/src/global/mixin/index.ts
@@ -5,7 +5,7 @@ import { defaultTheme } from '@global/theme';
 // eslint-disable-next-line import/prefer-default-export
 export const RoundScrollBar = css`
   &::-webkit-scrollbar {
-    width: 8px;
+    width: 6px;
   }
   &::-webkit-scrollbar-thumb {
     background-color: rgba(0, 0, 0, 0.4);
